// packages/evat/src/generated/types.ts
export interface BasicHeaderType {
  /**
   * Identifier of the request/response, unique with the taxnumber in every data
   * exchange transaction
   *
   * A kérés/válasz azonosítója, minden üzenetváltásnál - adószámonként - egyedi
   */
  requestId: string;
  /**
   * UTC time of the request/response
   *
   * A kérés/válasz keletkezésének UTC ideje
   */
  timestamp: string;
  /**
   * Request version number, indicating which datastructure the client sends data
   * in, and in which the response is expected
   *
   * A kérés/válasz verziószáma, hogy a hívó melyik interfész verzió szerint küld
   * adatot és várja a választ
   */
  requestVersion: string;
  /**
   * Header version number
   *
   * A header verziószáma
   */
  headerVersion?: string;
}
// ...
export interface CryptoType {
  /** Text content of the element. */
  value: string;
  cryptoType: string;
}
// ...
export interface ManageAttachmentUploadResponse {
  /**
   * Transactional data of the response
   *
   * A válasz tranzakcionális adatai
   */
  header: BasicHeaderType;
  /**
   * Basic result data
   *
   * Alap válaszeredmény adatok
   */
  result: BasicResultType;
  /**
   * Identifier received when uploading the attachment
   *
   * A melléklet feltöltésekor kapott azonosító
   */
  claimCheckId: string;
  /**
   * Start of validity of the upload identifier
   *
   * Feltöltési azonosító érvényességének kezdete
   */
  claimCheckIdValidFrom: string;
  /**
   * End of validity of the upload identifier
   *
   * Feltöltési azonosító érvényességének vége
   */
  claimCheckIdValidTo: string;
}
// ...
export interface ManageDeclarationFinalizeResponse {
  /**
   * Transactional data of the response
   *
   * A válasz tranzakcionális adatai
   */
  header: BasicHeaderType;
  /**
   * Basic result data
   *
   * Alap válaszeredmény adatok
   */
  result: BasicResultType;
  /**
   * Processing identifier
   *
   * Feldolgozási azonosító
   */
  declarationProcessingId: string;
}
// ...
export interface ManageDeclarationPartitionResponse {
  /**
   * Transactional data of the response
   *
   * A válasz tranzakcionális adatai
   */
  header: BasicHeaderType;
  /**
   * Basic result data
   *
   * Alap válaszeredmény adatok
   */
  result: BasicResultType;
  /**
   * Upload identifier
   *
   * Feltöltési azonosító
   */
  declarationUploadId: string;
  /**
   * Serial number of the uploaded partition
   *
   * A feltöltött partíció száma
   */
  partition: number;
}
// ...
export interface ManageDeclarationSubmissionResponse {
  /**
   * Transactional data of the response
   *
   * A válasz tranzakcionális adatai
   */
  header: BasicHeaderType;
  /**
   * Basic result data
   *
   * Alap válaszeredmény adatok
   */
  result: BasicResultType;
  /**
   * Name and value of the hash algorithm calculated for the uploaded content
   *
   * A feltöltött állományra számított hash algoritmus neve és értéke
   */
  contentHash: CryptoType;
  /**
   * Summary data of the declaration
   *
   * A bevallás összegző adatai
   */
  declarationSummary?: DeclarationSummaryType;
  /**
   * BEVFELD XML data based on VAT analytics
   *
   * Az ÁFA analitika alapján összeállított BEVFELD XML adatai
   */
  bevfeldData?: string;
}
// ...
export interface ManageDeclarationUploadResponse {
  /**
   * Transactional data of the response
   *
   * A válasz tranzakcionális adatai
   */
  header: BasicHeaderType;
  /**
   * Basic result data
   *
   * Alap válaszeredmény adatok
   */
  result: BasicResultType;
  /**
   * Upload identifier
   *
   * Feltöltési azonosító
   */
  declarationUploadId: string;
  /**
   * Start of validity of upload identifier
   *
   * Feltöltési azonosító érvényességének kezdete
   */
  declarationUploadValidFrom: string;
  /**
   * End of validity of upload identifier
   *
   * Feltöltési azonosító érvényességének vége
   */
  declarationUploadValidTo: string;
  /**
   * Identifier of the previous upload that was cancelled
   *
   * Az esetlegesen megszakított korábbi feltöltés azonosítója
   */
  cancelledDeclarationUploadId?: string;
}
// ...
export interface PurgeAttachmentResponse {
  /**
   * Transactional data of the response
   *
   * A válasz tranzakcionális adatai
   */
  header: BasicHeaderType;
  /**
   * Basic result data
   *
   * Alap válaszeredmény adatok
   */
  result: BasicResultType;
}
// ...
export interface QueryAttachmentListResponse {
  /**
   * Transactional data of the response
   *
   * A válasz tranzakcionális adatai
   */
  header: BasicHeaderType;
  /**
   * Basic result data
   *
   * Alap válaszeredmény adatok
   */
  result: BasicResultType;
  /**
   * List of uploaded attachments
   *
   * Feltöltött mellékletek listája
   */
  attachmentList?: AttachmentListType;
}
// ...
export interface QueryCustomsDeclarationDigestRequest {
  /**
   * Transactional data of the request
   *
   * A kérés tranzakcionális adatai
   */
  header: BasicHeaderType;
  /**
   * Authentication data of the request
   *
   * A kérés authentikációs adatai
   */
  user: UserHeaderType;
  /**
   * Management software data
   *
   * Az ügyviteli program adatai
   */
  software: SoftwareType;
  /**
   * The queried page count
   *
   * A lekérdezni kívánt lap száma
   */
  page: number;
  /**
   * Importer or indirect customs representative query parameter
   *
   * Importőri vagy közvetett vámjogi képviselői keresés paramétere
   */
  declarationDirection: DeclarationDirectionType;
  /**
   * Customs declaration query parameters
   *
   * Vámáru határozatok lekérdezési paraméterei
   */
  declarationQueryParams: DeclarationQueryType;
}
// ...
export interface QueryCustomsDeclarationDigestResponse {
  /**
   * Transactional data of the response
   *
   * A válasz tranzakcionális adatai
   */
  header: BasicHeaderType;
  /**
   * Basic result data
   *
   * Alap válaszeredmény adatok
   */
  result: BasicResultType;
  /**
   * The highest available page count matching the query
   *
   * A lekérdezés eredménye szerint elérhető utolsó lapszám
   */
  availablePage: number;
  /**
   * The currently queried page count
   *
   * A jelenleg lekérdezett lapszám
   */
  currentPage: number;
  /**
   * Customs declaration digest query result
   *
   * Vámáru határozat kivonat lekérdezési eredményei
   */
  declarationDigest?: DeclarationDigestType[];
}
// ...
export interface QueryCustomsDeclarationTaxCodeRequest {
  /**
   * Transactional data of the request
   *
   * A kérés tranzakcionális adatai
   */
  header: BasicHeaderType;
  /**
   * Authentication data of the request
   *
   * A kérés authentikációs adatai
   */
  user: UserHeaderType;
  /**
   * Management software data
   *
   * Az ügyviteli program adatai
   */
  software: SoftwareType;
  /**
   * Unique identifier of the customs declaration
   *
   * A vámáru határozat egyedi azonosítója
   */
  cdpsId: string;
  /**
   * Identifier of the resolution
   *
   * Határozatszám
   */
  resolutionId: string;
  /**
   * Importer or indirect customs representative query parameter
   *
   * Importőri vagy közvetett vámjogi képviselői keresés paramétere
   */
  declarationDirection: DeclarationDirectionType;
}
// ...
export interface QueryCustomsDeclarationTaxCodeResponse {
  /**
   * Transactional data of the response
   *
   * A válasz tranzakcionális adatai
   */
  header: BasicHeaderType;
  /**
   * Basic result data
   *
   * Alap válaszeredmény adatok
   */
  result: BasicResultType;
  /**
   * Customs declaration digest query result
   *
   * Vámáru határozat kivonat lekérdezési eredménye
   */
  declarationDigest?: DeclarationDigestType;
  /**
   * Tax code data
   *
   * Adókód adatok
   */
  taxCodeInformation?: TaxCodeInformationType[];
}
// ...
export interface QueryDeclarationListResponse {
  /**
   * Transactional data of the response
   *
   * A válasz tranzakcionális adatai
   */
  header: BasicHeaderType;
  /**
   * Basic result data
   *
   * Alap válaszeredmény adatok
   */
  result: BasicResultType;
  /**
   * List of declarations
   *
   * Bevallások listája
   */
  declarationList?: DeclarationListType;
  /**
   * List of statements
   *
   * Nyilatkozatok listája
   */
  statementList?: StatementListType;
}
// ...
export interface QueryDeclarationProcessingStatusResponse {
  /**
   * Transactional data of the response
   *
   * A válasz tranzakcionális adatai
   */
  header: BasicHeaderType;
  /**
   * Basic result data
   *
   * Alap válaszeredmény adatok
   */
  result: BasicResultType;
  /**
   * Declaration processing status
   *
   * Bevallás feldolgozási állapota
   */
  declarationProcessingStatus?: DeclarationProcessingStatusType;
}
// ...
export interface QueryDocumentListResponse {
  /**
   * Transactional data of the response
   *
   * A válasz tranzakcionális adatai
   */
  header: BasicHeaderType;
  /**
   * Basic result data
   *
   * Alap válaszeredmény adatok
   */
  result: BasicResultType;
  /**
   * Query identifier
   *
   * Lekérdezési azonosító
   */
  queryId: string;
}
// ...
export interface QueryDocumentListResultResponse {
  /**
   * Transactional data of the response
   *
   * A válasz tranzakcionális adatai
   */
  header: BasicHeaderType;
  /**
   * Basic result data
   *
   * Alap válaszeredmény adatok
   */
  result: BasicResultType;
  /**
   * Document list query result
   *
   * Bizonylat lista lekérdezési eredmények
   */
  documentList?: DocumentListType;
}
// ...
export interface QueryInvoiceTaxCodeRequest {
  /**
   * Transactional data of the request
   *
   * A kérés tranzakcionális adatai
   */
  header: BasicHeaderType;
  /**
   * Authentication data of the request
   *
   * A kérés authentikációs adatai
   */
  user: UserHeaderType;
  /**
   * Management software data
   *
   * Az ügyviteli program adatai
   */
  software: SoftwareType;
  /**
   * Invoice number param of the Invoice query
   *
   * Számla lekérdezés számlaszám paramétere
   */
  invoiceNumberQuery: InvoiceNumberQueryType;
  /**
   * Additional parameters for taxcoded invoice with VAT-relevant warings query
   *
   * Számla lekérdezés adókódokkal és ÁFA releváns warnokkal kiegészítő
   * paraméterei
   */
  additionalInvoiceTaxCodeQueryParams: AdditionalInvoiceTaxCodeQueryParamsType;
}
// ...
export interface QueryInvoiceTaxCodeResponse {
  /**
   * Transactional data of the response
   *
   * A válasz tranzakcionális adatai
   */
  header: BasicHeaderType;
  /**
   * Basic result data
   *
   * Alap válaszeredmény adatok
   */
  result: BasicResultType;
  /**
   * Invoice digest query result
   *
   * Számla kivonat lekérdezési eredmény
   */
  invoiceDigest?: InvoiceDigestType;
  /**
   * Tax code data
   *
   * Adókód adatok
   */
  taxCodeInformation?: TaxCodeInformationType[];
  /**
   * VAT-relevant warnings
   *
   * ÁFA szempontjából releváns figyelmeztetések
   */
  vatRelevantWarning?: VatRelevantWarningType[];
}
// ...
export interface QueryTaxCodeCatalogResponse {
  /**
   * Transactional data of the response
   *
   * A válasz tranzakcionális adatai
   */
  header: BasicHeaderType;
  /**
   * Basic result data
   *
   * Alap válaszeredmény adatok
   */
  result: BasicResultType;
  /**
   * Tax code catalogue data
   *
   * Adókód katalógus adatok
   */
  taxCodeCatalog?: TaxCodeCatalogType;
}
// ...
export interface UserHeaderType {
  /**
   * Login name of the technical user
   *
   * A technikai felhasználó login neve
   */
  login: string;
  /**
   * Hash value of the technical user's password
   *
   * A technikai felhasználó jelszavának hash értéke
   */
  passwordHash: CryptoType;
  /**
   * The taxpayer's tax number, whose name the technical user operates in
   *
   * A rendszerben regisztrált adózó adószáma, aki nevében a technikai
   * felhasználó tevékenykedik
   */
  taxNumber: string;
  /**
   * The taxpayer's tax number, whose name the technical user operates in
   *
   * A jogelőd adószáma, aki nevében a technikai felhasználó tevékenykedik
   */
  predecessorTaxNumber?: string;
  /**
   * Hash value of the request's signature
   *
   * A kérés aláírásának hash értéke
   */
  requestSignature: CryptoType;
}